# docs/api-servas-swagger.yaml
openapi: 3.2.0
info:
  title: API Reserva de Servicios
  description: Especificación alineada al plan de trabajo oficial y las Historias de Usuario (HU-001 a HU-040).
  version: 1.0.0
paths:
...
  /services/{id}/schedules:
    post:
      tags: [Agendas & Disponibilidad]
      summary: Configurar días y horarios de atención (HU-014, HU-015)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                required: [day_of_week, start_time, end_time]
                properties:
                  day_of_week: { type: integer, example: 2 } # 1=Lunes, 2=Martes...
                  start_time: { type: string, example: "09:00" }
                  end_time: { type: string, example: "16:00" }
      responses:
        '201': { description: Horarios de atención guardados. }

    put:
      tags: [Agendas & Disponibilidad]
      summary: Modificar días y horarios de atención (HU-016)
      description: Reemplaza la configuración de horarios existente por la nueva lista.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                required: [day_of_week, start_time, end_time]
                properties:
                  day_of_week: { type: integer, example: 2 } # 1=Lunes, 2=Martes...
                  start_time: { type: string, example: "09:00" }
                  end_time: { type: string, example: "16:00" }
      responses:
        '200': { description: Horarios de atención actualizados. }
  /services/{id}/availability:
    get:
      tags: [Agendas & Disponibilidad]
      summary: Calendario y franjas horarias disponibles (HU-018, HU-019, HU-020)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: date
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        '200': { description: Horarios disponibles calculados automáticamente tras evaluar bloqueos y reservas. }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT